requirements.txt
# PostgreSQL only. SQLite has been removed.

# AI-powered shipment document processor (OpenAI):
openai>=1.0.0       # OpenAI API for text/vision extraction
pdfplumber>=0.10.0  # PDF text extraction
openpyxl>=3.1.0     # Excel .xlsx support (required by pandas)
pandas>=2.0.0       # Excel/CSV and table structure
python-docx>=1.0.0  # Word document extraction
Pillow>=10.0.0      # Image handling
PyMuPDF>=1.24.0     # PDF → images for vision (optional, for PDF vision path)

# Optional: legacy document scraping
# pdfplumber, pandas, openpyxl also used by shipment processor above

# Web viewer dependencies:
Flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0

# Image matching dependencies (Deep Learning):
torch>=2.0.0        # PyTorch for deep learning
torchvision>=0.15.0 # Pre-trained models and image transforms
Pillow>=9.0.0       # Image processing
numpy>=1.21.0       # Numerical operations
scikit-learn>=1.0.0 # For cosine similarity (optional, we implement our own)

# Barcode scanning dependencies:
pyzbar>=0.1.9       # Barcode and QR code scanning
opencv-python>=4.5.0 # Image processing for barcode detection

# Calendar integration dependencies:
icalendar>=5.0.0    # iCal/Google Calendar support
pytz>=2023.3        # Timezone support

# Metadata extraction dependencies (FREE - no paid APIs):
spacy>=3.4.0        # Natural language processing (runs locally)
fuzzywuzzy>=0.18.0  # Fuzzy string matching
python-Levenshtein>=0.12.0  # Required by fuzzywuzzy for better performance
requests>=2.28.0    # For free barcode lookup APIs (Open Food Facts)
ollama>=0.1.0       # Optional: Local LLM for better category naming (hybrid approach)

# Receipt generation dependencies:
reportlab>=4.0.0   # PDF generation for receipts
qrcode>=7.4.0       # QR code generation for receipts
python-barcode>=0.15.0  # Barcode generation (Code128, Code39, etc.)
Pillow>=9.0.0       # Image processing (already listed above, but needed for qrcode and barcode)

# SMS CRM dependencies:
# Email-to-SMS uses built-in smtplib (no additional packages needed)
# boto3>=1.26.0     # Optional: Only needed when migrating to AWS SNS (~$0.006/SMS)

# Payment processing dependencies:
stripe>=7.0.0       # Stripe API for payment processing
cryptography>=41.0.0  # Encryption for storing sensitive API keys

# Clerk authentication dependencies:
clerk-backend-api>=1.0.0  # Clerk Backend SDK for user creation and invitations

# PostgreSQL database dependencies (local database):
psycopg2-binary>=2.9.0    # PostgreSQL adapter for Python
python-dotenv>=1.0.0      # Environment variable management
